// admin/src/generated/api/catalog/models/productPriceWindowDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { ProductPriceWindowDtoEndsAt } from './productPriceWindowDtoEndsAt';

export interface ProductPriceWindowDto {
  id: string;
  amount: string;
  startsAt: string;
  /** @nullable */
  endsAt?: ProductPriceWindowDtoEndsAt;
  status: string;
  /** @minimum 0 */
  version: number;
  createdAt: string;
}